extract abstract sql representer class
