+++ config/locales/en.yml
@@ -609,6 +609,10 @@ en:
         unremovable: "cannot be removed."
         wrong_length: "is the wrong length (should be %{count} characters)."
       models:
+        capability:
+          context:
+            global: 'Global'
+
         custom_field:
           at_least_one_custom_option: "At least one option needs to be available."
         custom_actions:
